ref: move getMockLocals to admin module
