+++ src/app/modules/form/admin-form/admin-form.service.ts
@@ -8,7 +8,7 @@ import {
   MAX_UPLOAD_FILE_SIZE,
   VALID_UPLOAD_FILE_TYPES,
 } from '../../../../shared/constants'
-import { DashboardFormView } from '../../../../types'
+import { AuthType, DashboardFormView, MyInfoAttribute } from '../../../../types'
 import getFormModel from '../../../models/form.server.model'
 import { DatabaseError } from '../../core/core.errors'
 import { MissingUserError } from '../../user/user.errors'
@@ -165,4 +165,31 @@ export const createPresignedPostForLogos = (
   InvalidFileTypeError | CreatePresignedUrlError
 > => {
   return createPresignedPost(AwsConfig.logoS3Bucket, uploadParams)
+}
+
+type SpcpLocals =
+  | {
+      uinFin: string
+      hashedFields: Set<MyInfoAttribute>
+    }
+  | { uinFin: string; userInfo: string }
+  | { [key: string]: never } // empty object
+export const getMockSpcpLocals = (
+  authType: AuthType,
+  myInfoAttrs: MyInfoAttribute[],
+): SpcpLocals => {
+  switch (authType) {
+    case AuthType.SP:
+      return {
+        uinFin: 'S1234567A',
+        hashedFields: new Set(myInfoAttrs),
+      }
+    case AuthType.CP:
+      return {
+        uinFin: '123456789A',
+        userInfo: 'ABC',
+      }
+    default:
+      return {}
+  }
 }
